refactor(engine): extract Scheduler to its own TU
Moves 13 step/prefill/decode methods (~1250 LOC, the bulk of engine.cpp)
to src/runtime/engine_scheduler.cpp:
  - step, step_async_graph_resume, step_schedule
  - supports_chunked_prefill_, resolve_prefill_chunk_size_
  - step_prefill, prefill_allocate_kv_blocks_,
    prefill_upload_metadata_, step_prefill_one
  - step_decode, decode_build_inference_state_,
    step_decode_forward, step_decode_process_outputs

All declarations in engine.h unchanged. Bodies byte-identical.

This is the largest single drop in Phase 4 of the architecture refactor.

Phase 4 of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
