+++ CMakeLists.txt
@@ -226,6 +226,7 @@ set(IMP_RUNTIME_SOURCES
     src/runtime/engine_kv_cache_init.cpp
     src/runtime/engine_workspace_warmup.cpp
     src/runtime/engine_graph_decode.cpp
+    src/runtime/engine_scheduler.cpp
     src/runtime/vision_pipeline.cpp
     src/runtime/constraint_manager.cpp
     src/runtime/vram_budget.cpp
